requirements.txt
dotenv==0.9.9
ffmpeg-python==0.2.0
openai==1.70.0
pydub==0.25.1
